requirements.txt
gradio
llama-cpp-python
requests
beautifulsoup4
googlesearch-python
